Golden Liquid Test Cleanup
